+++ gateway/packages/ui/src/auth/useJWT.ts
@@ -3,7 +3,7 @@ import { useState } from 'react'
 const LOCAL_STORAGE_JWT_KEY = 'jwt'
 
 export function useJWT(): [string | null, (token: string | null) => void] {
-  const item = localStorage.getItem(LOCAL_STORAGE_JWT_KEY)
+  const item = sessionStorage.getItem(LOCAL_STORAGE_JWT_KEY)
 
   const [token, setToken] = useState<null | string>(item)
 
@@ -11,9 +11,9 @@ export function useJWT(): [string | null, (token: string | null) => void] {
     token,
     (token) => {
       if (token === null) {
-        localStorage.removeItem(LOCAL_STORAGE_JWT_KEY)
+        sessionStorage.removeItem(LOCAL_STORAGE_JWT_KEY)
       } else {
-        localStorage.setItem(LOCAL_STORAGE_JWT_KEY, token)
+        sessionStorage.setItem(LOCAL_STORAGE_JWT_KEY, token)
       }
       setToken(token)
     },
